consolidacion_y_normalizacion/_codigo_fuente/requirements.txt
auto-py-to-exe==2.42.0
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
typing_extensions==4.9.0
Unidecode==1.3.8
